added contract_numbers to the rating table (u. h. to delete old tables)

+++ backend/src/main/java/edu/hsog/flat/backend/model/Rating.java
@@ -15,7 +15,7 @@ public class Rating {
 	private Long id;
 
 	@NonNull
-	private Integer ratingId;
+	private Long ratingId;
 
 	/*@NonNull
 	@OneToOne(targetEntity = Customer.class, fetch= FetchType.EAGER)
